requirements.txt
langchain
langchain-community
langchain-groq
langgraph
chromadb
sentence-transformers
streamlit
pypdf
python-dotenv